requirements.txt
requests==2.32.3
APScheduler==3.10.4
pytest==8.3.4
yoyo-migrations==8.2.0
